requirements.txt
pyserial
wx
